+++ PowerShell/Set-UserRightsAssignment/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Set-UserRightsAssignment release history
 
+## 2.0.0 - 2022-09-16
+
+* Added `SID` Support for `Principal`
+* Added Resolve-Principal support for both `PrincipalSID` and `PrincipalName`
+* Added random file names for all used files
+* Added some code optimizations
+* Replaced `Identity` with `Principal`
+* Fixed some variable declarations
+* Fixed `Replace` Support
+
+
 ## 1.0.1 - 2022-09-15
 
 * Added `Replace` option.